[C]: fixes for length.

+++ aeron-driver/src/main/c/agent/aeron_driver_agent.h
@@ -42,7 +42,7 @@ typedef struct aeron_driver_agent_frame_log_header_stct
     int64_t time_ms;
     int32_t result;
     int32_t sockaddr_len;
-    size_t message_len;
+    int32_t message_len;
 }
 aeron_driver_agent_frame_log_header_t;
